Moved models to its own package
