+++ app/src/main/java/br/com/ecostage/mobilecollect/model/Collect.kt
@@ -1,4 +1,4 @@
-package br.com.ecostage.mobilecollect.ui.collect
+package br.com.ecostage.mobilecollect.model
 
 import java.util.*
