+++ app/src/main/java/br/com/ecostage/mobilecollect/repository/CollectRepository.kt
@@ -1,7 +1,7 @@
 package br.com.ecostage.mobilecollect.repository
 
 import br.com.ecostage.mobilecollect.OnCollectLoadedListener
-import br.com.ecostage.mobilecollect.ui.collect.Collect
+import br.com.ecostage.mobilecollect.model.Collect
 import br.com.ecostage.mobilecollect.ui.collect.CollectInteractor
 
 /**
