+++ app/src/main/java/br/com/ecostage/mobilecollect/ui/collect/CollectInteractor.kt
@@ -1,5 +1,7 @@
 package br.com.ecostage.mobilecollect.ui.collect
 
+import br.com.ecostage.mobilecollect.model.Collect
+
 /**
  * Created by cmaia on 7/20/17.
  */
