+++ app/src/main/java/br/com/ecostage/mobilecollect/ui/profile/collect/CollectListAdapter.kt
@@ -9,7 +9,7 @@ import android.widget.BaseAdapter
 import android.widget.TextView
 import br.com.ecostage.mobilecollect.OnCollectLoadedListener
 import br.com.ecostage.mobilecollect.R
-import br.com.ecostage.mobilecollect.ui.collect.Collect
+import br.com.ecostage.mobilecollect.model.Collect
 import br.com.ecostage.mobilecollect.ui.collect.CollectActivity
 import br.com.ecostage.mobilecollect.ui.collect.CollectViewModel
 import org.jetbrains.anko.startActivity
